func: backward compatibility (blender 3.2.0)
Settled on blender 3.2.0

Despite the missing nodes it might be possible to retrograde to 3.0.0, but the geonodes are very buggy. I ran into all kinds of crashes when using "basic" geonodes functionnalities (realize instances, color attribute....). I even managed to corrupt some blend files just by attempting to debug the node tree.

So for now, 3.2.0 it is.

GN : imported the GN setup into blender 3.2.0 and fixed all the missing / disconnected nodes:
- switches
- compare nodes
- store named attribute

Also needed to fix the rotation for Transform Geometry nodes (it seems rotation is a vector in 3.2, and something else in 4.2 ?)

ADDON :
- added a get_geonode_inputs_from_modifier() to handle geonodes inputs access before 4.0.0 (node_group.inputs) and after 4.0.0 (node_group.interface.items.tree)
- moved the path to the asset file in funcs.py in the add_new_sgf_object() function